update overlay form
