+++ app/src/i18n/locales/en.json
@@ -75,6 +75,19 @@
 		"bg_border_width": "Border width",
 		"bg_border_opacity": "Border opacity",
 		"bg_border_radius": "Corner radius",
+		"group_placement": "Placement",
+		"group_placement_hint": "Where the overlay shows up and how big it is",
+		"group_timing": "Timing & sound",
+		"group_timing_hint": "How long it stays and how loud it plays",
+		"group_appearance": "Appearance",
+		"group_appearance_hint": "Media opacity, background, border",
+		"group_text": "Text",
+		"group_text_hint": "Styling applied to text messages",
+		"group_types": "Enabled media types",
+		"group_types_hint": "Which messages trigger the overlay",
+		"customize_background": "Customize background",
+		"customize_border": "Customize border",
+		"preview_summary": "{{size}}% · {{position}} · {{duration}}s · opacity {{opacity}}%",
 		"monitor": "Overlay screen",
 		"monitor_hint": "Select which screen the overlay appears on.",
 		"monitor_primary": "Primary",
