+++ app/src/i18n/locales/fr.json
@@ -75,6 +75,19 @@
 		"bg_border_width": "Épaisseur de bordure",
 		"bg_border_opacity": "Opacité de bordure",
 		"bg_border_radius": "Arrondi des angles",
+		"group_placement": "Placement",
+		"group_placement_hint": "Où s'affiche l'overlay et à quelle taille",
+		"group_timing": "Durée & son",
+		"group_timing_hint": "Combien de temps il reste et à quel volume",
+		"group_appearance": "Apparence",
+		"group_appearance_hint": "Opacité du média, arrière-plan, bordure",
+		"group_text": "Texte",
+		"group_text_hint": "Mise en forme des messages textes",
+		"group_types": "Types de médias activés",
+		"group_types_hint": "Quels messages déclenchent l'affichage",
+		"customize_background": "Personnaliser l'arrière-plan",
+		"customize_border": "Personnaliser la bordure",
+		"preview_summary": "{{size}}% · {{position}} · {{duration}}s · opacité {{opacity}}%",
 		"monitor": "Écran de l'overlay",
 		"monitor_hint": "Choisissez sur quel écran l'overlay s'affiche.",
 		"monitor_primary": "Principal",
