+++ app/src/windows/settings/components/overlay-form/components/section-header.tsx
@@ -1,0 +1,18 @@
+import { Separator } from "@memeover/ui/components/ui/separator";
+
+interface SectionHeaderProps {
+	title: string;
+	hint?: string;
+}
+
+export function SectionHeader({ title, hint }: SectionHeaderProps) {
+	return (
+		<div className="space-y-3">
+			<div className="space-y-1">
+				<h2 className="font-display text-base tracking-wide">{title}</h2>
+				{hint && <p className="text-xs text-muted-foreground">{hint}</p>}
+			</div>
+			<Separator />
+		</div>
+	);
+}
