+++ app/src/windows/settings/components/overlay-form/overlay-form.tsx
@@ -1,11 +1,14 @@
 import { NbCard } from "@memeover/ui/components/branded/nb-card";
 import type { Settings } from "@/shared/types";
 import { FullscreenCompatibilityNotice } from "./fullscreen-compatibility-notice";
-import { BackgroundSettingsSection } from "./sections/background-settings-section";
-import { DisplaySettingsSection } from "./sections/display-settings-section";
+import { AppearanceSection } from "./sections/appearance-section";
 import { FormActions } from "./sections/form-actions";
 import { MediaTypesSection } from "./sections/media-types-section";
 import { OverlayControls } from "./sections/overlay-controls";
+import { PlacementSection } from "./sections/placement-section";
+import { StickyPreview } from "./sections/sticky-preview";
+import { TextSection } from "./sections/text-section";
+import { TimingSoundSection } from "./sections/timing-sound-section";
 import { useOverlayForm } from "./use-overlay-form";
 
 export interface OverlayFormProps {
@@ -17,33 +20,58 @@ export function OverlayForm({ initialData }: OverlayFormProps) {
 
 	return (
 		<div className="p-5">
-			<div className="mx-auto max-w-2xl space-y-5">
+			<div className="mx-auto max-w-4xl space-y-5">
 				{/* ── Fullscreen compatibility notice ── */}
 				<FullscreenCompatibilityNotice />
 
-				{/* ── Overlay Controls ── */}
+				{/* ── Overlay Controls (orthogonal to form state) ── */}
 				<NbCard>
 					<OverlayControls />
 				</NbCard>
 
-				{/* ── Display & Background & Media Settings ── */}
-				<NbCard>
-					<form
-						onSubmit={(e) => {
-							e.preventDefault();
-							void form.handleSubmit();
-						}}
-					>
-						<form.AppForm>
-							<div className="space-y-5">
-								<DisplaySettingsSection />
-								<BackgroundSettingsSection />
-								<MediaTypesSection />
+				{/* ── Settings form ── */}
+				<form
+					onSubmit={(e) => {
+						e.preventDefault();
+						void form.handleSubmit();
+					}}
+				>
+					<form.AppForm>
+						<div className="grid gap-5 md:grid-cols-[minmax(0,1fr)_280px]">
+							{/* Main column — sections grouped by user intent */}
+							<div className="space-y-5 min-w-0">
+								<NbCard>
+									<PlacementSection />
+								</NbCard>
+
+								<NbCard>
+									<TimingSoundSection />
+								</NbCard>
+
+								<NbCard>
+									<AppearanceSection />
+								</NbCard>
+
+								<NbCard>
+									<TextSection />
+								</NbCard>
+
+								<NbCard>
+									<MediaTypesSection />
+								</NbCard>
+
 								<FormActions isPending={isPending} />
 							</div>
-						</form.AppForm>
-					</form>
-				</NbCard>
+
+							{/* Sticky preview sidebar — desktop only */}
+							<aside className="hidden md:block md:sticky md:top-5 md:self-start">
+								<NbCard>
+									<StickyPreview />
+								</NbCard>
+							</aside>
+						</div>
+					</form.AppForm>
+				</form>
 			</div>
 		</div>
 	);
